vitalglobal_com/requirementsnew.txt
arrow==1.2.2
asgiref==3.5.2
attrs==21.4.0
beautifulsoup4==4.11.1
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.1.0
cryptography==37.0.4
DateTime==4.5
Django==4.0.6
django-recaptcha==3.0.0
gunicorn==20.1.0
ics==0.7.2
idna==3.3
mysql-client==0.0.1
pycparser==2.21
PyMySQL==1.0.2
python-dateutil==2.8.2
pytz==2022.1
requests==2.28.1
six==1.16.0
soupsieve==2.3.2.post1
sqlparse==0.4.2
TatSu==5.8.3
tzdata==2022.1
urllib3==1.26.11
whitenoise==6.2.0
zope.interface==5.4.0
